chore(ai): doc js/services/playbackService.js
