+++ context/CONTEXT_2026-02-13_12-31-20.md
@@ -1,0 +1,13 @@
+# Docs Code Investigator Context
+
+- **Date:** 2026-02-13
+- **Agent:** docs-code-investigator
+- **Target File:** `js/services/playbackService.js`
+- **Line Count:** ~630 LOC (based on read_file output, though `wc` said 1288 earlier - maybe I read a different version or `wc` counts comments/blanks differently? Let me check `wc` again later. Ah, the `read_file` output seems complete but maybe shorter than I thought. Wait, scrolling up... the file content I read ends with `export default PlaybackService;`. It seems complete. 600 lines is still substantial.)
+- **Reason:** Core component for Hybrid Playback Strategy, complex state machine in `PlaybackSession`, essential for app reliability. Listed in `AGENTS.md` as a key file for Playback subsystem.
+- **Plan:**
+    1. Analyze `js/services/playbackService.js`.
+    2. Add JSDoc to `PlaybackService` and `PlaybackSession`.
+    3. Document the `execute()` flow with inline comments.
+    4. Create `docs/playbackService-overview.md` detailing the internal state machine and probe logic.
+    5. Run lint and unit tests.
