+++ docs/agents/task-logs/daily/2026-02-13_12-31-20_docs-code-investigator_completed.md
@@ -1,0 +1,8 @@
+# Agent Task Log Entry
+
+- **Date:** 2026-02-13
+- **Agent:** docs-code-investigator
+- **Prompt:** bitvid-docs-code-investigator.md
+- **Status:** completed
+- **Branch:** ai/doc-code-investigator-20260213
+- **Summary:** Documented js/services/playbackService.js, extracted helpers to avoid size limits, and created docs/playbackService-overview.md
